requirements.txt
coverage==7.3.2
Django==3.2.10
djangorestframework==3.12.4
drf-spectacular>=0.22.1,<0.23
django-filter==23.2
factory-boy==3.3.0
Faker==19.3.1
gunicorn==20.1.0
hupper==1.12
psycopg2==2.9.5
psycopg2-binary==2.9.1
